[Non-working Checkpoint for Slate-Plugins 1.0.0 migration] May have finally cracked the code
